% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeaheadsa.R
\name{my_new_function}
\alias{my_new_function}
\title{Multiply by two}
\usage{
my_new_function(x)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
A numeric value, double the input.
}
\description{
This function takes a numeric input and returns double the value.
}
